hotzone/requirements.txt
asgiref==3.3.0
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
dj-email-url==1.0.1
Django==3.1.3
django-cache-url==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
environs==9.0.0
gunicorn==20.0.4
idna==2.10
isort==5.6.4
joblib==0.17.0
lazy-object-proxy==1.4.3
marshmallow==3.9.0
mccabe==0.6.1
numpy==1.19.4
psycopg2-binary==2.8.6
pylint==2.6.0
python-dotenv==0.15.0
pytz==2020.4
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.2
urllib3==1.25.11
whitenoise==5.2.0
wrapt==1.12.1
